src/bot/ml/form_detector.py
from typing import Dict, List, Optional, Tuple
import numpy as np
from selenium.webdriver.common.by import By
from selenium.webdriver.remote.webelement import WebElement
from bs4 import BeautifulSoup
import tensorflow as tf
import re
import logging

class FormField:
    def __init__(self, element: WebElement, field_type: str, name: str, required: bool):
        self.element = element
        self.field_type = field_type
        self.name = name
        self.required = required
        self.confidence_score = 0.0
        self.mapped_value = None

class FormDetector:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        
        # Common field patterns
        self.field_patterns = {
            'name': r'(?i)(name|full.?name|first.?name|last.?name)',
            'email': r'(?i)(email|e-mail)',
            'phone': r'(?i)(phone|mobile|cell)',
            'experience': r'(?i)(experience|years.?of.?experience)',
            'education': r'(?i)(education|degree|qualification)',
            'resume': r'(?i)(resume|cv|upload.?resume|attach.?cv)',
            'cover_letter': r'(?i)(cover.?letter|letter.?of.?interest)',
            'linkedin': r'(?i)(linkedin|linkedin.?profile)',
            'portfolio': r'(?i)(portfolio|work.?samples)',
            'salary': r'(?i)(salary|expected.?salary|salary.?requirement)'
        }
        
        # Load pre-trained model for field classification
        self.model = self._load_model()

    def _load_model(self) -> Optional[tf.keras.Model]:
        """
        Load pre-trained model for field classification
        """
        try:
            # TODO: Replace with actual model path
            return tf.keras.models.load_model('path/to/form_detection_model')
        except Exception as e:
            self.logger.error(f"Failed to load model: {e}")
            return None

    async def detect_fields(self, driver) -> List[FormField]:
        """
        Detect and classify form fields on the page
        """
        try:
            # Get all potential form elements
            form_elements = self._get_form_elements(driver)
            
            # Analyze and classify each field
            detected_fields = []
            
            for element in form_elements:
                field_info = await self._analyze_field(element)
                if field_info:
                    detected_fields.append(field_info)
            
            # Post-process and validate detected fields
            return self._post_process_fields(detected_fields)
            
        except Exception as e:
            self.logger.error(f"Field detection failed: {e}")
            return []

    def _get_form_elements(self, driver) -> List[WebElement]:
        """
        Get all potential form input elements
        """
        elements = []
        
        # Common input selectors
        selectors = [
            "input[type='text']",
            "input[type='email']",
            "input[type='tel']",
            "input[type='file']",
            "textarea",
            "select",
            ".form-control",
            "[role='textbox']"
        ]
        
        for selector in selectors:
            elements.extend(driver.find_elements(By.CSS_SELECTOR, selector))
            
        return elements

    async def _analyze_field(self, element: WebElement) -> Optional[FormField]:
        """
        Analyze and classify a single form field
        """
        try:
            # Get field attributes
            attributes = self._get_element_attributes(element)
            
            # Get surrounding context
            context = self._get_field_context(element)
            
            # Determine field type
            field_type, confidence = await self._classify_field(attributes, context)
            
            if field_type and confidence > 0.7:  # Confidence threshold
                return FormField(
                    element=element,
                    field_type=field_type,
                    name=attributes.get('name', ''),
                    required=self._is_field_required(element, attributes, context)
                )
                
            return None
            
        except Exception as e:
            self.logger.debug(f"Field analysis failed: {e}")
            return None

    def _get_element_attributes(self, element: WebElement) -> Dict:
        """
        Extract relevant attributes from element
        """
        attributes = {}
        important_attrs = ['name', 'id', 'class', 'placeholder', 'aria-label', 'type']
        
        for attr in important_attrs:
            try:
                value = element.get_attribute(attr)
                if value:
                    attributes[attr] = value
            except:
                continue
                
        return attributes

    def _get_field_context(self, element: WebElement) -> str:
        """
        Get surrounding text context for the field
        """
        try:
            # Get label if exists
            label = self._find_label(element)
            
            # Get nearby text
            nearby_text = self._get_nearby_text(element)
            
            return f"{label} {nearby_text}".strip()
            
        except Exception as e:
            self.logger.debug(f"Context extraction failed: {e}")
            return ""

    async def _classify_field(self, attributes: Dict, context: str) -> Tuple[str, float]:
        """
        Classify field type using ML model and pattern matching
        """
        # Pattern-based classification
        for field_type, pattern in self.field_patterns.items():
            if (re.search(pattern, context) or 
                any(re.search(pattern, str(v)) for v in attributes.values())):
                return field_type, 0.9
        
        # ML-based classification if available
        if self.model:
            try:
                # Prepare input features
                features = self._prepare_features(attributes, context)
                
                # Get model prediction
                prediction = self.model.predict(features)
                
                # Get highest confidence prediction
                field_type_idx = np.argmax(prediction)
                confidence = prediction[0][field_type_idx]
                
                return self._idx_to_field_type(field_type_idx), float(confidence)
                
            except Exception as e:
                self.logger.error(f"ML classification failed: {e}")
        
        return None, 0.0

    def _is_field_required(self, element: WebElement, attributes: Dict, context: str) -> bool:
        """
        Determine if field is required
        """
        # Check required attribute
        if element.get_attribute('required'):
            return True
            
        # Check for required indicators in context
        required_patterns = [
            r'(?i)required',
            r'(?i)mandatory',
            r'\*'
        ]
        
        for pattern in required_patterns:
            if (re.search(pattern, context) or 
                any(re.search(pattern, str(v)) for v in attributes.values())):
                return True
                
        return False

    def _find_label(self, element: WebElement) -> str:
        """
        Find associated label text for element
        """
        try:
            # Check for label element
            label_id = element.get_attribute('aria-labelledby')
            if label_id:
                label_elem = element.find_element(By.ID, label_id)
                if label_elem:
                    return label_elem.text
            
            # Check for aria-label
            aria_label = element.get_attribute('aria-label')
            if aria_label:
                return aria_label
            
            # Check for nearby label
            parent = element.find_element(By.XPATH, '..')
            labels = parent.find_elements(By.TAG_NAME, 'label')
            if labels:
                return labels[0].text
                
        except Exception:
            pass
            
        return ""

    def _post_process_fields(self, fields: List[FormField]) -> List[FormField]:
        """
        Post-process and validate detected fields
        """
        # Remove duplicates
        unique_fields = {}
        for field in fields:
            if field.field_type not in unique_fields or \
               field.confidence_score > unique_fields[field.field_type].confidence_score:
                unique_fields[field.field_type] = field
        
        # Validate required fields presence
        required_types = ['name', 'email', 'resume']
        for req_type in required_types:
            if req_type not in unique_fields:
                self.logger.warning(f"Required field type missing: {req_type}")
        
        return list(unique_fields.values())

    def _prepare_features(self, attributes: Dict, context: str) -> np.ndarray:
        """
        Prepare features for ML model
        """
        # TODO: Implement feature extraction for ML model
        pass

    def _idx_to_field_type(self, idx: int) -> str:
        """
        Convert model output index to field type
        """
        # TODO: Implement mapping from model output to field type
        pass
